36kWc - Add HTML rendering
